refactor custom config functions
